Add files via upload
This data set includes various biomedical voice measurements from a total of 31 individuals, 23 of whom have been identified as having Parkinson's disease. The table displays a distinct voice measurement in every column, linked to among the 195 voice recordings from the individuals listed in the "name" column in each row. The information in the "status column" is meant to differentiate between individuals who are healthy and those who have PD, with a value of 0 indicating a healthy individual and a value of 1 indicating a PD case. The information is in CSV format encoded in ASCII. The CSV file contains rows, each representing a single voice recording instance. Each patient has approximately six recordings, with the patient's name listed in the 1st column.